+++ test/test.c
@@ -7,17 +7,16 @@
 void
 show_tensor(Tensor(float) *tensor)
 {
-	float * array = tensor_byte_data((Tensor*)tensor);
+	float * array = tensor_byte_data(TENSOR(tensor));
 
 	printf("Tensor {");
 
-	for(size_t i = 0; i < tensor_elements((Tensor*)tensor); i ++)
+	for(size_t i = 0; i < tensor_elements(TENSOR(tensor)); i ++)
 	{
 		if(i == 0)
 			printf("%f", array[i]);
 		else
 			printf(", %f", array[i]);
-
 	}
 
 	printf("}\n");
@@ -31,8 +30,8 @@ main(void)
 	fflush(stdout);
 
 	Tensor(float) * tensor = 
-		tensor_float_new_random(3, (size_t[]){1,2,2});
-
+		tensor_float_new_random(3, (size_t[]) {1,2,2});
+ 
 	if(tensor != NULL)
 	{
 		printf("Tensor successfully allocated.\n");
@@ -43,7 +42,7 @@ main(void)
 		printf("Releasing tensor\n");
 		fflush(stdout);
 
-		tensor_finalize((Tensor*) tensor);
+		tensor_delete(TENSOR(tensor));
 
 		printf("Program exit successfully.\n");
 		fflush(stdout);